+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * table multiple=false 时单选框勾选状态不显示 ([#961](https://github.com/WeBankFinTech/fes-design/issues/961)) ([46d6f24](https://github.com/WeBankFinTech/fes-design/commit/46d6f24c9dab2d3d41ffe028969a5be64ea1965c))
 * vite8 css 变量处理问题 ([#963](https://github.com/WeBankFinTech/fes-design/issues/963)) ([df12c80](https://github.com/WeBankFinTech/fes-design/commit/df12c800130920d9cc1991677fc3a05d6b625037))
+* **layout:** FSider 样式兼容 (box-sizing: border-box) ([#887](https://github.com/WeBankFinTech/fes-design/issues/887))
 
 
 ### Features